new and improved timer
The timer is now calculated as a delta between a fixed point and the
current system time. Time remaining is displayed to a precision of a
hundredth of a second. reset() has been brought back from the dead and
new timer functions have been written to support the new timer
functionality.